service-b-python/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
prometheus-client==0.19.0